formatted all balances with their respective currencies - lifted to store
