+++ packages/nextjs/app/lending/components/YourBorrows.tsx
@@ -5,6 +5,7 @@ import useAccountAddress from "@/hooks/useAccount";
 import useGetReservesData from "@/hooks/useGetReservesData";
 import useGetUserReservesData from "@/hooks/useGetUserReservesData";
 import { useTotalBalance } from "@/hooks/useTotalBalance";
+import { useLendingStore } from "@/stores/lending-store";
 import { Address } from "viem";
 import { useTranslation } from "~~/app/context/LanguageContext";
 import { useTotalAPY } from "~~/hooks/useTotalAPY";
@@ -22,6 +23,7 @@ const YourBorrows: React.FC<YourBorrowsProps> = ({ setBorrowsTotalBalance }) => 
   const { reservesData, isLoading: isLoadingReserves, isError: isErrorReserves } = useGetReservesData();
   const { userReservesData, isLoading: isLoadingUserReserves, isError: isErrorUserReserves } = useGetUserReservesData();
   const { address: walletAddress } = useAccountAddress();
+  const { formatBalanceWithCurrency } = useLendingStore();
 
   const [balances, setBalances] = useState<Record<string, string>>({});
   const [reservesWithBalances, setReservesWithBalances] = useState<any[]>([]);
@@ -130,6 +132,7 @@ const YourBorrows: React.FC<YourBorrowsProps> = ({ setBorrowsTotalBalance }) => 
                     tokenAddress={reserve.variableDebtTokenAddress as Address}
                     walletAddress={walletAddress as Address}
                     onBalanceChange={handleVariableDebtChange}
+                    formatDisplay={balance => formatBalanceWithCurrency(balance, reserve.symbol)}
                   />
                 </p>
               </div>
